requirements.txt
crewai>=0.28.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
pyyaml>=6.0
